+++ Gruntfile.js
@@ -1,4 +1,5 @@
 // Require external packages.
+const autoprefixer = require( 'autoprefixer' );
 const sass = require( 'node-sass' );
 
 // Set up a versioned folder for SassDoc
@@ -145,7 +146,36 @@ module.exports = ( grunt ) => {
 				},
 				files: {
 					'docs/css/docs.css': '_docs/css-dev/docs.scss',
+					'css/burf-base.css': 'css-dev/burf-base.scss',
+					'css/burf-theme.css': 'css-dev/burf-theme.scss',
 				},
+			},
+		},
+		postcss: {
+			defaults: {
+				options: {
+					map: {
+						inline: false, // Save all sourcemaps as separate files.
+						annotation: 'css/', // Save to this specified directory.
+					},
+					processors: [
+						autoprefixer, // add vendor prefixes.
+					],
+				},
+				src: ['css/burf-theme.css', 'css/burf-base.css'],
+			},
+			docs: {
+				options: {
+					map: {
+						inline: false, // Save all sourcemaps as separate files.
+						annotation: 'docs/css/', // Save to this specified directory.
+					},
+					processors: [
+						autoprefixer, // add vendor prefixes.
+					],
+				},
+				src: 'docs/css/docs.css',
+				dest: 'docs/css/docs.css',
 			},
 		},
 		watch: {
@@ -164,8 +194,11 @@ module.exports = ( grunt ) => {
 				tasks: [ 'js', 'concat' ],
 			},
 			styles: {
-				files: [ '_docs/css-dev/*.scss', 'css-dev/**/*.scss' ],
-				tasks: [ 'sass' ],
+				files: [
+					'_docs/css-dev/*.scss',
+					'css-dev/**/*.scss',
+				],
+				tasks: ['sass', 'postcss'],
 			},
 			vendor: {
 				files: [ '_docs/vendor/**/*' ],
@@ -209,6 +242,7 @@ module.exports = ( grunt ) => {
 	} );
 
 	// Load Plugins.
+	grunt.loadNpmTasks( 'grunt-babel' );
 	grunt.loadNpmTasks( 'grunt-browser-sync' );
 	grunt.loadNpmTasks( 'grunt-browserify' );
 	grunt.loadNpmTasks( 'grunt-contrib-clean' );
@@ -217,10 +251,10 @@ module.exports = ( grunt ) => {
 	grunt.loadNpmTasks( 'grunt-contrib-uglify' );
 	grunt.loadNpmTasks( 'grunt-contrib-watch' );
 	grunt.loadNpmTasks( 'grunt-gh-pages' );
+  grunt.loadNpmTasks( 'grunt-postcss' );
 	grunt.loadNpmTasks( 'grunt-sass' );
 	grunt.loadNpmTasks( 'grunt-sass-lint' );
 	grunt.loadNpmTasks( 'grunt-sassdoc' );
-	grunt.loadNpmTasks( 'grunt-babel' );
 
 	grunt.registerTask( 'build', [ 'js', 'sassdoc' ] );
 	grunt.registerTask( 'deploy', [ 'build', 'gh-pages' ] );